+++ build.properties
@@ -124,7 +124,6 @@
     #    apps/app-builder,\
     #    apps/application-list,\
     #    apps/archived,\
-    #    apps/archived-modules-upgrade,\
     #    apps/asset,\
     #    apps/batch-engine,\
     #    apps/bean-portlet,\
@@ -176,6 +175,7 @@
     #    apps/journal,\
     #    apps/knowledge-base,\
     #    apps/layout,\
+    #    apps/legacy-data-cleanup,\
     #    apps/license-manager,\
     #    apps/login,\
     #    apps/map,\
