+++ modules/apps/legacy-data-cleanup/legacy-data-cleanup-test/bnd.bnd
@@ -1,3 +1,3 @@
-Bundle-Name: Liferay Archived Modules Upgrade Test
-Bundle-SymbolicName: com.liferay.archived.modules.upgrade.test
+Bundle-Name: Liferay Legacy Data Cleanup Test
+Bundle-SymbolicName: com.liferay.legacy.data.cleanup.test
 Bundle-Version: 1.0.0
