+++ modules/apps/legacy-data-cleanup/legacy-data-cleanup-test/src/testIntegration/resources/com/liferay/legacy/data/cleanup/test/dependencies/private-messaging-tables.sql
