+++ modules/apps/legacy-data-cleanup/legacy-data-cleanup/src/main/java/com/liferay/legacy/data/cleanup/internal/upgrade/CleanupChat.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.archived.modules.upgrade.internal;
+package com.liferay.legacy.data.cleanup.internal.upgrade;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
