+++ modules/apps/legacy-data-cleanup/legacy-data-cleanup/src/main/java/com/liferay/legacy/data/cleanup/internal/upgrade/CleanupDictionary.java
@@ -12,23 +12,19 @@
  * details.
  */
 
-package com.liferay.archived.modules.upgrade.internal;
+package com.liferay.legacy.data.cleanup.internal.upgrade;
 
 /**
  * @author Sam Ziemer
  */
-public class CleanupDirectory extends BaseCleanupNoninstanceablePortlet {
+public class CleanupDictionary extends BaseCleanupNoninstanceablePortlet {
 
 	@Override
 	protected void doUpgrade() throws Exception {
 		removePortlet(
-			"com.liferay.directory.web",
-			new String[] {"11", "186", "187", "188"},
+			"com.liferay.dictionary.web", new String[] {"23"},
 			new String[] {
-				"com_liferay_directory_web_portlet_DirectoryPortlet",
-				"com_liferay_directory_web_portlet_FriendsDirectoryPortlet",
-				"com_liferay_directory_web_portlet_SiteMembersDirectoryPortlet",
-				"com_liferay_directory_web_portlet_MySitesDirectoryPortlet"
+				"com_liferay_dictionary_web_portlet_DictionaryPortlet"
 			});
 	}
